fixing our first big bug
All of our responses were coming back with single quotes which is
invalid for json.  This was because the set_data functionality
was not invoked.  so "data" was set to a dictionary.  Then a simple
str(response.data) is called internally and it just gave pythons
representation of a dict, which has single quotes.

This made ALL error responses invalid.  Good deal.