Fixing another broken case
Since it is in multiple places I plan to fix this in the better way.
Maybe just use flask.jsonify since I don't have any extra functionality
to add yet

Maybe find a way to make response.data use a __str__ so it can be
converted to json.  Maybe add a step somewhere at the end that will
convert the data to a string myself.

This should be a temporary... I hope