Allowing no trailing slash
This change is to allow you to query a directory with or without
a trailing slash on the resource.

BAD THING: It appears that getting of a "directory" exists when
going through boto/real aws and does NOT work through boto.

In other words

get_key("/lol/") does not exist through moto (as it does through
boto) even though we have an actual artifact "/lol/hi"