making test_base specific to bucket_update
This is starting to push how many levels of inheritance we should
have.  Basically, it was easy to do it this way and I am worried
about running out of time.

The real fix would likely be to break out this functionality into a
separate helper object but that could take some time